cree las modificaciones en la navBar para que se modifiquen segun login, artisan o comon, y logout, modifique main, para q aparezca siempre(yano hace falta importarla)
