+++ client/src/features/auth/components/navBar.jsx
@@ -1,0 +1,124 @@
+import { React, useState } from "react";
+import { Box, IconButton, Flex, Button, Link, Icon, Group } from "@chakra-ui/react";
+import { IoMenu } from "react-icons/io5";
+import { FaHandHoldingHand } from "react-icons/fa6";
+import ContactForm from "../../../shared/components/forms/contact";
+import LoginForm from "../../../shared/components/forms/login";
+import RegisterForm from "../../../shared/components/forms/register";
+import { useAuth } from "./../../../app/providers/AuthProvider";
+
+
+function NavBar() {
+const { userType, logout } = useAuth();
+const [isOpen, setIsOpen] = useState(false);
+const [showContactForm, setShowContactForm] = useState(false);
+const [showLoginForm, setShowLoginForm] = useState(false);
+const [showRegisterForm, setShowRegisterForm] = useState(false);
+
+const handleToggle = () => {
+    console.log("Botón clickeado");
+    setIsOpen( prevState => !prevState);
+};
+//formulario de contacto
+const handleShowContactForm = () => {
+    setShowContactForm(true); 
+};
+const handleCloseContactForm = () => {
+    setShowContactForm(false); 
+};
+//formulario login
+const handleShowLoginForm = () => {
+    setShowLoginForm(true); 
+};
+const handleCloseLoginForm = () => {
+    setShowLoginForm(false); 
+};
+//formulario de registro
+const handleShowRegisterForm = () => {
+    setShowRegisterForm(true);
+};
+const handleCloseRegisterForm = () => {
+    setShowRegisterForm(false); 
+};   
+
+return (
+    <Box bg="neutral" color="accent" p="4">
+        {/* Tamaño PC */}
+        <Flex direction="row" display={{ base: "none", md: "flex" }} justify="space-between">
+            <Flex gap="4" justify="flex-start">
+            <Icon fontSize="4xl">
+                <FaHandHoldingHand />
+            </Icon>
+            </Flex>
+            <Flex direction="row" gap="6" justify="flex-end">
+                <Link to="">Artesanos</Link>
+                <Link to="">Sobre Nosotros</Link>
+                <Link to="" onClick={handleShowContactForm}>Contáctanos</Link>
+                {userType === "comon" &&(
+                    <>
+                        <Link to="">Carrito</Link>
+                        <Link to="">Mis favoritos</Link>
+                    </>
+                )}
+                {userType === "artisan" && (
+                    <>
+                        <Link to="">Mis ventas</Link>
+                    </>
+                )}
+                {userType ? (
+                    <Button bg="secondary" onClick={logout}>Cerrar Sesión</Button>
+                ) : (
+                  <>
+                    <Button bg="secondary" onClick={handleShowLoginForm}>Ingresar</Button>
+                    <Button bg="primary" onClick={handleShowRegisterForm}>Registrarse</Button>
+                  </>
+                )}
+            </Flex>
+        </Flex>
+        <IconButton
+            aria-label="Abrir menú"
+            align="center"
+            textAlign="center"
+            display={{ base: "block", md: "none" }}
+            onClick={handleToggle}
+            bg="secondary"
+        >
+            <IoMenu fontSize="5xl"/>
+        </IconButton>
+
+        {/* Menu desplegable para pantallas pequeñas */}
+        {isOpen && (
+            <Flex direction="column" gap="2" align="center" >
+                <Link p="2" href="#">Artesanos</Link>
+                <Link p="2" href="#">Sobre Nosotros</Link>
+                <Link p="2" href="#" onClick={handleShowContactForm}>Contáctanos</Link>
+                {userType === "comon" &&(
+                    <>
+                        <Link to="">Carrito</Link>
+                        <Link to="">Mis favoritos</Link>
+                    </>
+                )}
+                {userType === "artisan" && (
+                    <>
+                        <Link to="">Mis ventas</Link>
+                    </>
+                )}
+                {userType ? (
+                    <Button bg="secondary" onClick={logout}>Cerrar Sesión</Button>
+                ) : (
+                    <>
+                        <Button bg="secondary" onClick={handleShowLoginForm}>Ingresar</Button>
+                        <Button bg="primary" onClick={handleShowRegisterForm}>Registrarse</Button>
+                    </>
+                )}
+            </Flex>
+        )}
+        {showContactForm && <ContactForm onClose={handleCloseContactForm} />}
+        {showLoginForm && <LoginForm onClose={handleCloseLoginForm} />}
+        {showRegisterForm && <RegisterForm onClose={handleCloseRegisterForm} />}
+    </Box>
+
+);
+}
+
+export default NavBar;
