+++ client/src/features/home/components/hero.jsx
