+++ client/src/features/home/pages/home.jsx
@@ -1,17 +1,15 @@
 import React from "react";
-import NavBar from "../../../shared/components/navBar";
 import Hero from "../components/hero";
 import CardConteinerPlus from "../components/cardConteinerPlus";
 import ArtisanConteiner from "../components/artisanConteiner";
 
-function Auth () {
+function Home () {
     return(
         <>
-            <NavBar />
             <Hero />
             <CardConteinerPlus />
             <ArtisanConteiner />
         </>
     );
 }
-export default Auth;
+export default Home;
