+++ client/src/features/profile/pages/Profile.jsx
@@ -1,4 +1,4 @@
-import NavBar from "../../../shared/components/navBar";
+import NavBar from "../../auth/components/navBar";
 import { AddProduct } from "../components/AddProduct";
 import { BannerProfile } from "../components/BannerProfile";
 import { CardList } from "../components/CardList";
